fix: platform-scope native storage modules and split-bundle ownership
