+++ apps/mobile/scripts/unionBuildHelpers.js
@@ -101,10 +101,17 @@ function buildRuntimeOwnership({
   );
 
   // Expand shared startup set with sync dependencies.
-  // A shared module (e.g., defiUtils.ts) may sync-depend on a module that
-  // only exists in one graph (e.g., a crypto lib only in the bg graph).
-  // That dep must also be in the common bundle, so promote it to shared.
-  // We follow sync deps in BOTH graphs to cover all transitive deps.
+  // A shared module (e.g., defiUtils.ts) may sync-depend on a module that is
+  // not yet in the shared startup set; that dep must also be loadable before
+  // the common bundle finishes evaluating, so promote it to shared.
+  //
+  // Only shared-equivalent deps may be promoted. The common bundle is
+  // serialized from the MAIN graph, so promoting a runtime-divergent dep
+  // (e.g. react-native-mmkv resolves to the real package in bg but to the
+  // main-thread guard shim in main) would either ship the main variant to the
+  // background runtime or drop bg-only modules from every bundle. Divergent
+  // deps stay runtime-owned: each runtime's eager bundle carries its own
+  // variant under the same stable module id.
   const pendingShared = [...sharedStartupAbsPaths];
   while (pendingShared.length > 0) {
     const current = pendingShared.pop();
@@ -115,7 +122,8 @@ function buildRuntimeOwnership({
         for (const [, dep] of mod.dependencies) {
           if (
             dep.data?.data?.asyncType !== 'async' &&
-            !sharedStartupAbsPaths.has(dep.absolutePath)
+            !sharedStartupAbsPaths.has(dep.absolutePath) &&
+            sharedEquivalentAbsPaths.has(dep.absolutePath)
           ) {
             // Promote this sync dep to shared
             sharedStartupAbsPaths.add(dep.absolutePath);
