+++ packages/kit-bg/src/states/jotai/jotaiStorageConsts.ts
@@ -1,0 +1,1 @@
+export const MMKV_MIGRATION_COMPLETE_KEY = '__mmkv_migration_v1__';
